refactor(frontend): single BitcoinTapTab entry point
Remove the re-export shim; import the screen from bitcoin/BitcoinTapTab
everywhere.
