--- dsm_client/frontend/src/components/screens/BitcoinTapTab.tsx
@@ -1,3 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Re-export from decomposed bitcoin/ directory.
-export { default } from './bitcoin/BitcoinTapTab';
